Reorganize Mods.Cnc
