+++ OpenRA.Mods.Cnc/Traits/Buildings/ProductionAirdrop.cs
@@ -15,7 +15,7 @@ using OpenRA.Mods.RA.Air;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Cnc
+namespace OpenRA.Mods.Cnc.Traits
 {
 	[Desc("Deliver the unit in production via skylift.")]
 	public class ProductionAirdropInfo : ProductionInfo
