+++ OpenRA.Mods.Cnc/Traits/Render/RenderGunboat.cs
