+++ OpenRA.Mods.Cnc/Traits/Render/WithDeliveryAnimation.cs
