+++ OpenRA.Mods.Cnc/Traits/SpawnViceroid.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Cnc
+namespace OpenRA.Mods.Cnc.Traits
 {
 	class SpawnViceroidInfo : ITraitInfo
 	{
